# openapi/openapi.yml
openapi: 3.1.0
info:
  title: MLB Forecast API
  description: "\n    # MLB Predictions Backend API\n\n    Enterprise-grade MLB game\
    \ prediction platform delivering real-time analytics and machine learning insights.\n\
    \n    ## \U0001F680 Core Features\n\n    ### \u26BE **Data Ingestion Pipeline**\n\
    \    - Real-time game results from official MLB APIs\n    - Historical statistics\
    \ and performance metrics\n    - Live game schedules and league standings\n  \
    \  - Automated data validation and cleansing\n\n    ### \U0001F916 **Advanced\
    \ Machine Learning**\n    - Probabilistic game outcome predictions\n    - Total\
    \ runs and score forecasting\n    - Player performance analytics\n    - Continuous\
    \ model training and optimization\n\n    ### \u26A1 **High-Performance Architecture**\n\
    \    - Multi-layer Redis caching system\n    - Optimized database queries with\
    \ indexing\n    - Asynchronous processing for scalability\n    - Sub-second API\
    \ response times\n\n    ### \U0001F4CA **Analytics & Insights**\n    - Team performance\
    \ trending\n    - Head-to-head historical analysis\n    - Weather impact on game\
    \ outcomes\n    - Injury reports integration\n\n    ## \U0001F527 Technical Stack\n\
    \n    - **Framework**: FastAPI with async/await patterns\n    - **Database**:\
    \ PostgreSQL with optimized schemas\n    - **Cache**: Redis for high-speed data\
...
    \    ```\n    "
  version: 1.0.0
paths:
...
  /api/v1/games:
    get:
      tags:
        - Games
      summary: List Games
      description: "Retrieve games with optional filtering and pagination.\n\nArgs:\n\
        \    date: Filter by date in YYYY-MM-DD format\n    team_id: Filter by team\
        \ ID\n    status: Filter by game status\n    limit: Maximum number of games\
        \ to return\n    use_cases: Dependency injected use cases\n\nReturns:\n  \
        \  JSONResponse: Standardized response with games list\n\nRaises:\n    DomainExceptions.InvalidDataError:\
        \ If filters are invalid"
      operationId: list_games_api_v1_games_get
      parameters:
        - name: date
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            description: Filter by date in YYYY-MM-DD format
            title: Date
          description: Filter by date in YYYY-MM-DD format
        - name: team_id
          in: query
          required: false
          schema:
            anyOf:
              - type: integer
              - type: 'null'
            description: Filter by team ID
            title: Team Id
          description: Filter by team ID
        - name: status
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            description: Filter by game status (scheduled, in_progress, completed)
            title: Status
          description: Filter by game status (scheduled, in_progress, completed)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            maximum: 200
            description: Maximum number of games to return
            default: 50
            title: Limit
          description: Maximum number of games to return
      responses:
        '200':
          description: Successfully retrieved games list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/games/{game_id}:
    get:
      tags:
        - Games
      summary: Get Game
      description: "Get a game by its ID.\n\nArgs:\n    game_id: The ID of the game\
        \ to retrieve\n    use_cases: Dependency injected use cases\n\nReturns:\n\
        \    JSONResponse: Standardized response with game details\n\nRaises:\n  \
        \  DomainExceptions.GameNotFoundError: If game is not found\n    DomainExceptions.InvalidDataError:\
        \ If game_id is invalid"
      operationId: get_game_api_v1_games__game_id__get
      parameters:
        - name: game_id
          in: path
          required: true
          schema:
            type: integer
            description: The ID of the game to get
            title: Game Id
          description: The ID of the game to get
      responses:
        '200':
          description: Successfully retrieved game details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDetailResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
  /api/v1/data/ingest/games:
    post:
      tags:
        - Games
        - Data Ingestion
      summary: Ingest Games
      description: "Ingest game data from external MLB API.\n\nArgs:\n    date: Specific\
        \ date to ingest games for in YYYY-MM-DD format\n    days_back: Number of\
        \ days back to ingest (max 30)\n    use_cases: Dependency injected use cases\n\
        \nReturns:\n    JSONResponse: Standardized response with ingestion results\n\
        \nRaises:\n    DomainExceptions.InvalidDataError: If date format is invalid\n\
        \    DomainExceptions.ExternalServiceError: If MLB API is unavailable"
      operationId: ingest_games_api_v1_data_ingest_games_post
      parameters:
        - name: date
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            description: Specific date to ingest games for (YYYY-MM-DD)
            title: Date
          description: Specific date to ingest games for (YYYY-MM-DD)
        - name: days_back
          in: query
          required: false
          schema:
            type: integer
            maximum: 30
            description: Number of days back to ingest games for
            default: 7
            title: Days Back
          description: Number of days back to ingest games for
      responses:
        '201':
          description: Successfully ingested game data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataIngestionResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
                description: Bad Request
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
                description: Not Found
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnprocessableEntity'
                description: Unprocessable Entity
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalServerError'
                description: Internal Server Error
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceUnavailable'
                description: Service Unavailable
...
components:
  schemas:
...
    DataIngestionResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          $ref: '#/components/schemas/IngestionSummaryData'
          description: Ingestion operation details
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          description: Success message
      type: object
      required:
        - code
        - data
        - message
      title: DataIngestionResponse
      description: Response model for data ingestion endpoints.
      content:
        application/json:
          example:
            status: success
            code: 201
            data:
              ingestion_summary:
                operation: team_ingestion
                records_processed: 30
                records_created: 30
                records_updated: 0
                errors: [ ]
                duration_seconds: 2.45
                timestamp: '2025-07-25T18:30:15.123456Z'
              sample_teams:
                - id: 1
                  mlb_id: 147
                  name: New York Yankees
                  abbreviation: NYY
                  city: New York
                  division: American League East
                  league: American League
                  venue_name: Yankee Stadium
                  created_at: '2025-07-25T17:58:28.647662Z'
            errors: [ ]
            message: Successfully ingested 30 teams
    DataIngestionResultDTO:
      properties:
        operation:
          type: string
          title: Operation
          description: Type of ingestion operation
        records_processed:
          type: integer
          title: Records Processed
          description: Number of records processed
        records_created:
          type: integer
          title: Records Created
          description: Number of new records created
        records_updated:
          type: integer
          title: Records Updated
          description: Number of existing records updated
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: Any errors encountered
        duration_seconds:
          type: number
          title: Duration Seconds
          description: Operation duration in seconds
        timestamp:
          type: string
          format: date-time
          title: Timestamp
          description: Operation timestamp
      type: object
      required:
        - operation
        - records_processed
        - records_created
        - records_updated
        - duration_seconds
        - timestamp
      title: DataIngestionResultDTO
      description: Data Transfer Object for data ingestion operation results.
    BadRequest:
      description: Bad Request - Invalid input parameters
      title: Bad Request
      $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found - Resource not found
      title: Not Found
      $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Validation Error - Request validation failed
      title: Validation Error
      $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal Server Error
      title: Internal Server Error
      $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: Service Unavailable - External service error
      title: Service Unavailable
      $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      properties:
        status:
          type: string
          title: Status
          description: Response status
          default: error
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          type: 'null'
          title: Data
          description: No data in error responses
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          anyOf:
            - type: string
            - type: 'null'
          title: Message
          description: Error message
      type: object
      required:
        - code
        - errors
      title: ErrorResponse
      description: Error response model for API documentation.
      examples:
        - code: 400
          errors:
            - League must be 'American' or 'National'
          message: Invalid data provided
          status: error
        - code: 404
          errors:
            - Team with ID 999 not found
          message: Resource not found
          status: error
        - code: 500
          errors:
            - An unexpected error occurred
          message: Internal server error
          status: error
...
    GameDTO:
      properties:
        id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Id
          description: Internal game ID
        mlb_game_id:
          type: integer
          title: Mlb Game Id
          description: Official MLB game ID
        home_team_id:
          type: integer
          title: Home Team Id
          description: Home team ID
        away_team_id:
          type: integer
          title: Away Team Id
          description: Away team ID
        game_date:
          type: string
          format: date-time
          title: Game Date
          description: Game date and time
        status:
          type: string
          title: Status
          description: Game status
        scheduled_innings:
          type: integer
          title: Scheduled Innings
          description: Scheduled innings
          default: 9
        home_score:
          anyOf:
            - type: integer
            - type: 'null'
          title: Home Score
          description: Home team final score
        away_score:
          anyOf:
            - type: integer
            - type: 'null'
          title: Away Score
          description: Away team final score
        winning_team_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Winning Team Id
          description: Winning team ID
        created_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Created At
          description: Record creation timestamp
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
          description: Record last update timestamp
      type: object
      required:
        - mlb_game_id
        - home_team_id
        - away_team_id
        - game_date
        - status
      title: GameDTO
      description: Data Transfer Object for Game entity.
      example:
        id: 1
        mlb_game_id: 12345
        home_team_id: 147
        away_team_id: 121
        game_date: '2025-07-25T19:05:00Z'
        status: completed
        scheduled_innings: 9
        home_score: 7
        away_score: 4
        winning_team_id: 147
        created_at: '2025-07-25T17:58:28.647662Z'
        updated_at: '2025-07-25T22:30:15.123456Z'
    GameDetailResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          $ref: '#/components/schemas/GameDTO'
          description: Game details
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          description: Success message
      type: object
      required:
        - code
        - data
        - message
      title: GameDetailResponse
      description: Response model for single game endpoints.
      example:
        status: success
        code: 200
        data:
          id: 1
          mlb_game_id: 12345
          home_team_id: 147
          away_team_id: 121
          game_date: '2025-07-25T19:05:00Z'
          status: completed
          scheduled_innings: 9
          home_score: 7
          away_score: 4
          winning_team_id: 147
          created_at: '2025-07-25T17:58:28.647662Z'
          updated_at: '2025-07-25T22:30:15.123456Z'
        errors: [ ]
        message: Game 12345 retrieved successfully
    GameListResponse:
      properties:
        status:
          type: string
          title: Status
          default: success
        code:
          type: integer
          title: Code
          description: HTTP status code
        data:
          items:
            $ref: '#/components/schemas/GameDTO'
          type: array
          title: Data
          description: List of games
        errors:
          items:
            type: string
          type: array
          title: Errors
          description: List of error messages
        message:
          type: string
          title: Message
          description: Success message
      type: object
      required:
        - code
        - data
        - message
      title: GameListResponse
      description: Response model for game list endpoints.
      example:
        status: success
        code: 200
        data:
          - id: 1
            mlb_game_id: 12345
            home_team_id: 147
            away_team_id: 121
            game_date: '2025-07-25T19:05:00Z'
            status: scheduled
            scheduled_innings: 9
            created_at: '2025-07-25T17:58:28.647662Z'
        errors: [ ]
        message: Retrieved 15 games successfully
...
    IngestionSummaryData:
      properties:
        ingestion_summary:
          $ref: '#/components/schemas/DataIngestionResultDTO'
          description: Ingestion operation summary
        sample_teams:
          anyOf:
            - items:
                $ref: '#/components/schemas/TeamDTO'
              type: array
            - type: 'null'
          title: Sample Teams
          description: Sample of ingested teams
        sample_games:
          anyOf:
            - items:
                $ref: '#/components/schemas/GameDTO'
              type: array
            - type: 'null'
          title: Sample Games
          description: Sample of ingested games
      type: object
      required:
        - ingestion_summary
      title: IngestionSummaryData
      description: Data model for ingestion operation responses.
...
    TeamDTO:
      properties:
        id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Id
          description: Internal team ID
        mlb_id:
          type: integer
          title: Mlb Id
          description: Official MLB team ID
        name:
          type: string
          title: Name
          description: Team name
        abbreviation:
          type: string
          title: Abbreviation
          description: Team abbreviation (e.g., NYY)
        city:
          type: string
          title: City
          description: Team city
        division:
          type: string
          title: Division
          description: Division (East, Central, West)
        league:
          type: string
          title: League
          description: League (American, National)
        venue_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Venue Name
          description: Home venue name
        created_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Created At
          description: Record creation timestamp
        updated_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Updated At
          description: Record last update timestamp
      type: object
      required:
        - mlb_id
        - name
        - abbreviation
        - city
        - division
        - league
      title: TeamDTO
      description: Data Transfer Object for Team entity.
...